pyproject.toml
[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[project]
name = "modalflow"
version = "0.1.0"
description = "Serverless Airflow Executor on Modal"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=3.0.6",
    "modal>=1.3.0",
    "click>=8.0.0",
]

[project.scripts]
modalflow = "modalflow.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock",
    "black",
    "ruff",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
